test(deployer): fix missing rwx vm cleanup after flag refactor
In a previous PR the composer flags were silenced in favour of a composer wrapper
which cleaned up by itself (without using composer).

This introduced a regression where the rwx vm was not getting killed after the test
was complete.
This may affect the subsequent tests.

To tackle this we move the wrapper to the deployer itself, and let the deployer
query the wrapper for cleanup.
